Add player the ability to collect items
In this commit I add collect logic. As the result players can collect items.
Dogs now have the bag and can collect or remove items from it.
Dog's bag capacity sets using config file.

+++ src/session/player.h
@@ -21,7 +21,7 @@ public:
         return dog_.GetName();
     }
 
-    const uint16_t& GetDogID() const {
+    const size_t& GetDogID() const {
         return dog_.GetID();
     }
